refactor: move core silent-payments functions to a core directory
