+++ src/core/encoding.ts
