+++ src/core/interface.ts
