+++ src/core/outputs.ts
