+++ src/core/scanning.ts
