+++ test/utility.spec.ts
@@ -1,5 +1,5 @@
-import { Outpoint, PrivateKey } from '../src/interface';
-import { calculateSumOfPrivateKeys, hashOutpoints } from '../src';
+import { Outpoint, PrivateKey } from '../src/core/interface';
+import { calculateSumOfPrivateKeys, hashOutpoints } from '../src/core';
 import { inputPrivateKeys, outpoints } from './fixtures/utility';
 
 describe('Utility', () => {
